[PHY] Revert LR-FHSS bw back to `uint8_t`

+++ src/protocols/PhysicalLayer/PhysicalLayer.h
@@ -58,7 +58,7 @@ struct FSKRate_t {
 */
 struct LrFhssRate_t {
   /*! \brief Bandwidth */
-  float bw;
+  uint8_t bw;
 
   /*! \brief Coding rate */
   uint8_t cr;
